Huge overhaul, now passing check
Some functions removed, renamed or changed, so that the package is not
entirely backwards compatible. The remaining functions accept both phylo
and phylo4 objects, and are documented. The functions nConfig,
ladderSizes and getDepths (on which most of the others depend) have been
freshly implemented by Caroline Colijn and are significantly faster.